[FIX] fix building semantics

+++ network/cvae_graph/dataloader.py
@@ -45,7 +45,7 @@ class GraphDataset(Dataset):
             node_features = torch.tensor(np.array([graph.nodes[node]['node_features'] for node in graph.nodes()]),
                                          dtype=torch.float32)
             node_semantics = torch.tensor(np.array([graph.nodes[node]['node_semantics'] for node in graph.nodes()]),
-                                          dtype=torch.float32)
+                                          dtype=torch.long)
             building_masks = torch.tensor(np.array([graph.nodes[node]['building_masks'] for node in graph.nodes()]),
                                           dtype=torch.float32)
